s16 to s8 change for yaw rate

+++ c/include/libsbp/ssr.h
@@ -126,7 +126,7 @@ generating configuration
   u8 mw_consistency;     /**< Consistency indicator for Melbourne-Wubbena linear combinations
  */
   u16 yaw;                /**< Satellite yaw angle [1 / 256 semi-circle] */
-  s16 yaw_rate;           /**< Satellite yaw angle rate [1 / 8192 semi-circle / s] */
+  s8 yaw_rate;           /**< Satellite yaw angle rate [1 / 8192 semi-circle / s] */
   phase_biases_content_t biases[8];          /**< Phase biases corrections for a
 satellite being tracked.
  */
